NIFI-7200: Revert "NIFI-6530 - HTTP SiteToSite server returns 201 in case no data is available"
This reverts commit f01668e66ad2e45197915769e966a4be27e1592e.

+++ nifi-commons/nifi-site-to-site-client/src/main/java/org/apache/nifi/remote/client/socket/SocketClient.java
@@ -23,7 +23,6 @@ import org.apache.nifi.remote.TransactionCompletion;
 import org.apache.nifi.remote.TransferDirection;
 import org.apache.nifi.remote.client.AbstractSiteToSiteClient;
 import org.apache.nifi.remote.client.SiteToSiteClientConfig;
-import org.apache.nifi.remote.exception.NoContentException;
 import org.apache.nifi.remote.protocol.DataPacket;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -126,13 +125,14 @@ public class SocketClient extends AbstractSiteToSiteClient {
         }
 
         final EndpointConnection connectionState = pool.getEndpointConnection(direction, getConfig());
+        if (connectionState == null) {
+            return null;
+        }
 
         final Transaction transaction;
         try {
             transaction = connectionState.getSocketClientProtocol().startTransaction(
                     connectionState.getPeer(), connectionState.getCodec(), direction);
-        } catch (final NoContentException e) {
-            return null;
         } catch (final Throwable t) {
             pool.terminate(connectionState);
             throw new IOException("Unable to create Transaction to communicate with " + connectionState.getPeer(), t);
